+++ src/app/_apis/auth/user-type/user-type.ts
@@ -1,6 +1,6 @@
 import { END_POINTS } from "@/constants/api";
 import { UserTypeData, UserTypeResponse } from "app/_types/user-type";
-import { apiGet, apiPost } from "app/apis/methods";
+import { apiGet, apiPost } from "app/_apis/methods";
 
 export const postUserType = (body?: UserTypeData) => {
   return apiPost(END_POINTS.POST_USER_TYPE, body);
